# api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/auth.login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [
          {
            "name": "x-forwarded-for",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "user-agent",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "access_token": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": ""
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "security": [
          {}
        ],
        "summary": "Login with email and password",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth.forgotPassword": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [
          {
            "name": "x-forwarded-for",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reset email sent (if email exists)",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "security": [
          {}
        ],
        "summary": "Request password reset email",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth.resetPassword": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [
          {
            "name": "x-forwarded-for",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successful",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": ""
          },
          "400": {
            "description": "Invalid or expired token"
          }
        },
        "security": [
          {}
        ],
        "summary": "Reset password with token",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth.sessions": {
      "get": {
        "operationId": "AuthController_sessions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of active sessions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "user_id": {
                        "type": "string"
                      },
                      "ip_address": {
                        "type": "string",
                        "nullable": true
                      },
                      "user_agent": {
                        "type": "string",
                        "nullable": true
                      },
                      "expires_at": {
                        "type": "string"
                      },
                      "revoked_at": {
                        "type": "string",
                        "nullable": true
                      },
                      "created_at": {
                        "type": "string"
                      },
                      "updated_at": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "List active sessions",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth.revokeSession": {
      "post": {
        "operationId": "AuthController_revokeSession",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session revoked",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Revoke a specific session",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth.revokeAllSessions": {
      "post": {
        "operationId": "AuthController_revokeAllSessions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "All sessions revoked",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Logout from all devices",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/track": {
      "post": {
        "operationId": "EventsController_track",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackEventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {}
        ],
        "summary": "Track a single event",
        "tags": [
          "events"
        ]
      }
    },
    "/api/track.batch": {
      "post": {
        "operationId": "EventsController_trackBatch",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackBatchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {}
        ],
        "summary": "Track multiple events in a batch",
        "tags": [
          "events"
        ]
      }
    },
...
  },
  "info": {
    "title": "Staminads API",
    "description": "Web analytics platform for tracking TimeScore metrics",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "jwt-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      },
      "demo-secret": {
        "type": "apiKey",
        "in": "query",
        "name": "secret",
        "description": "Demo secret for demo endpoints"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "newPassword": {
            "type": "string",
            "minLength": 8,
            "maxLength": 72
          }
        },
        "required": [
          "token",
          "newPassword"
        ]
      },
...
      "TrackEventDto": {
        "type": "object",
        "properties": {
          "workspace_id": {
            "type": "string"
          },
          "session_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "duration": {
            "type": "number",
            "minimum": 0
          },
          "referrer": {
            "type": "string"
          },
          "referrer_domain": {
            "type": "string"
          },
          "referrer_path": {
            "type": "string"
          },
          "is_direct": {
            "type": "boolean"
          },
          "landing_page": {
            "type": "string"
          },
          "landing_domain": {
            "type": "string"
          },
          "landing_path": {
            "type": "string"
          },
          "utm_source": {
            "type": "string"
          },
          "utm_medium": {
            "type": "string"
          },
          "utm_campaign": {
            "type": "string"
          },
          "utm_term": {
            "type": "string"
          },
          "utm_content": {
            "type": "string"
          },
          "utm_id": {
            "type": "string"
          },
          "utm_id_from": {
            "type": "string"
          },
          "screen_width": {
            "type": "number"
          },
          "screen_height": {
            "type": "number"
          },
          "viewport_width": {
            "type": "number"
          },
          "viewport_height": {
            "type": "number"
          },
          "device": {
            "type": "string"
          },
          "browser": {
            "type": "string"
          },
          "browser_type": {
            "type": "string"
          },
          "os": {
            "type": "string"
          },
          "user_agent": {
            "type": "string"
          },
          "connection_type": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          },
          "max_scroll": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "sdk_version": {
            "type": "string"
          },
          "created_at": {
            "type": "number"
          },
          "updated_at": {
            "type": "number"
          },
          "properties": {
            "type": "object"
          }
        },
        "required": [
          "workspace_id",
          "session_id",
          "name",
          "path",
          "landing_page",
          "created_at",
          "updated_at"
        ]
      },
      "TrackBatchDto": {
        "type": "object",
        "properties": {
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TrackEventDto"
            }
          }
        },
        "required": [
          "events"
        ]
      },
...
    }
  }
}